Moved schema-related files to own folder
