+++ src/schemas/schema.webExtensionMsg.json
