+++ src/schemas/schema.weightedVariations.json
